Update to support joint bank accounts
Modified main classes to support the creation and management of joint bank accounts. In the 'Account' class, changed the single 'Client' object to a 'List' of 'Client' objects. Created a new 'split' method in the 'BankService' class that splits the joint account's balance equally between the owners. This update was necessary to meet the new requirement of allowing joint bank accounts and ensuring fair division of funds when a shared account is closed.